Add Update page class
[MAILPOET-2200]
